+++ apps/app/src/components/ScheduleCard/ScheduleCard.module.scss
@@ -188,7 +188,7 @@
       }
 
       .logo {
-        height: 12px;
+        height: 38px;
       }
     }
   }
