+++ apps/app/src/router/Router.tsx
@@ -8,7 +8,6 @@ import {
 import { RouteNames } from './routes';
 import { CompaniesPage } from '../pages/CompaniesPage';
 import { FlyTalksPage } from '../pages/FlyTalksPage';
-import { HomePage } from '../pages/HomePage';
 import { LoginPage } from '../pages/LoginPage';
 import { NotFoundPage } from '../pages/NotFoundPage';
 import { PasswordResetPage } from '../pages/PasswordResetPage';
@@ -17,6 +16,8 @@ import { RegisterPage } from '../pages/RegisterPage';
 import { SchedulePage } from '../pages/SchedulePage';
 import { ShoppingPage } from '../pages/ShoppingPage';
 import { NavigationLayout } from '../layout';
+import Home from '../pages/Home';
+import TestPage from '../pages/TestPage/TestPage';
 
 const router = createBrowserRouter(
   createRoutesFromElements(
@@ -25,13 +26,14 @@ const router = createBrowserRouter(
       <Route path={RouteNames.REGISTER} element={<RegisterPage />} />
       <Route path={RouteNames.PASSWORD_RESET} element={<PasswordResetPage />} />
       <Route element={<NavigationLayout />} errorElement={<>error</>}>
-        <Route path={RouteNames.HOME} element={<HomePage />} />
+        <Route path={RouteNames.HOME} element={<Home />} />
         <Route path={RouteNames.PROFILE} element={<ProfilePage />} />
         <Route path={RouteNames.COMPANIES} element={<CompaniesPage />} />
         <Route path={RouteNames.SCHEDULE} element={<SchedulePage />} />
         <Route path={RouteNames.FLY_TALKS} element={<FlyTalksPage />} />
         <Route path={RouteNames.SHOPPING} element={<ShoppingPage />} />
       </Route>
+      <Route path='/app/test' element={<TestPage />} />
       <Route path='*' element={<NotFoundPage />} />
     </React.Fragment>,
   ),
